Add infrastructure code for resume redirect
This commit adds the necessary Terraform code to create the infrastructure for a resume redirect. It includes the creation of an ACM certificate, Route 53 record, CloudFront distribution, S3 bucket, and associated policies and functions. The code is organized into separate files for each component, such as `acm.tf`, `cloudfront.tf`, `s3.tf`, etc. The `deploy.sh` script is also included to automate the deployment process. The infrastructure is designed to redirect the resume domain to a Google Docs URL specified in the `redirectFunction.js` file.